+++ src/TitleDetails/TitleDetailsActions/TitleDetailsExpectedActions.js
@@ -6,13 +6,19 @@ import {
   Button,
 } from '@folio/stripes/components';
 
-export function TitleDetailsExpectedActions({ openAddPieceModal, openReceiveList, hasReceive }) {
+export function TitleDetailsExpectedActions({
+  openAddPieceModal,
+  openReceiveList,
+  hasReceive,
+  disabled,
+}) {
   return (
     <>
       {hasReceive && (
         <Button
           data-test-title-receive-button
           onClick={openReceiveList}
+          disabled={disabled}
         >
           <FormattedMessage id="ui-receiving.title.details.button.receive" />
         </Button>
@@ -20,6 +26,7 @@ export function TitleDetailsExpectedActions({ openAddPieceModal, openReceiveList
       <Button
         data-test-add-piece-button
         onClick={openAddPieceModal}
+        disabled={disabled}
       >
         <FormattedMessage id="ui-receiving.piece.button.addPiece" />
       </Button>
@@ -31,4 +38,5 @@ TitleDetailsExpectedActions.propTypes = {
   hasReceive: PropTypes.bool.isRequired,
   openAddPieceModal: PropTypes.func.isRequired,
   openReceiveList: PropTypes.func.isRequired,
+  disabled: PropTypes.bool,
 };